PHPs PDO class is broke
So PHP has got some very strange habits like representing true as 1 and
0 as nothing, but the problem with the PDO class is backwards it works
fine when I try to set a bit value to false but when I try to set it to
true I get a SQL error saying the value provided is too long. So my
guess is that's it's turning the true into the string true or something
weird like that. But as is always the case there is a hack workaround